Refactored DevfilePageProjects
