+++ src/util/client/index.ts
@@ -1,2 +1,3 @@
+export * from './downloadHelperFuncs';
 export * from './getCSSStyles';
 export * from './stringFuncs';
